requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
aiogram>=3.0.0,<4.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0

